nav bar is now responsive
